requirements.txt
psycopg2
# confluent_kafka
kafka-python